refactor(Net_Sockets, Net_Interfaces): modernize network layer code
- socket_stream: Replace C-style cast with static_cast for strlen,
  initialize sockaddr struct with brace init, initialize peer_len at declaration
- socket_stream_reader: Replace C-style casts with static_cast
- socket_tls_tlsparameters: Replace C-style casts with reinterpret_cast,
  initialize SSL_CTX/SSL structs with brace init
- socket_udp: Replace C-style casts with reinterpret_cast, use std::make_shared
  instead of raw new, remove unused sys/types.h include
- socket_unix: Initialize sockaddr_un structs with brace init,
  replace C-style casts with reinterpret_cast, initialize server_len at declaration
- streams_bridge_thread: Remove unnecessary (uint16_t) casts where type already matches,
  initialize timeval with brace init
- Net_Interfaces: Update sAppExecCmd references to ExecutionArgs after struct rename

+++ Mantids30/Net_Sockets/streams_bridge_thread.cpp
@@ -25,7 +25,7 @@ void Bridge_Thread::setSocketEndpoints(const std::shared_ptr<Socket_Stream> &src
 bool Bridge_Thread::sendPing()
 {
     std::lock_guard<std::mutex> lock(mt_fwd);
-    bool r = m_dstSocket->writeU<uint16_t>((uint16_t) 0);
+    bool r = m_dstSocket->writeU<uint16_t>(0);
     return r;
 }
 
@@ -90,7 +90,7 @@ int Bridge_Thread::processPipe(Side fwd)
                 std::lock_guard<std::mutex> lock(mt_fwd);
 
                 // Write the size to be written (chunk)
-                if (!m_dstSocket->writeU<uint16_t>((uint16_t) bytesReceived))
+                if (!m_dstSocket->writeU<uint16_t>(bytesReceived))
                 {
                     return -2;
                 }
@@ -126,13 +126,13 @@ int Bridge_Thread::processPipe(Side fwd)
             }
 
             // Attempt to read from the dst
-            if (!m_dstSocket->readFull(curBlock, (uint16_t) bytesReceived))
+            if (!m_dstSocket->readFull(curBlock, bytesReceived))
             {
                 return -1;
             }
 
             // Attempt to write to the src
-            if (!src->writeFull(curBlock, (uint16_t) bytesReceived))
+            if (!src->writeFull(curBlock, bytesReceived))
             {
                 return -2;
             }
